+++ src/Loader/Messages/Exeptions.js
@@ -1,42 +1,26 @@
 const ansis = require('ansis');
-const { bgRed, yellow, cyan, green, ansi256, cyanBright } = require('ansis/colors');
+const { yellow, cyan, green, ansi256, cyanBright, reset, whiteBright } = require('ansis/colors');
 const { pluginName } = require('../../config');
 
 const redBright = ansi256(203);
-
 const pluginHeaderHtml = `<span style="color:#e36049">[${pluginName}]</span>`;
-const pluginHeader = `\n${bgRed.whiteBright` ${pluginName} `}`;
-let lastError = null;
 
 class LoaderException extends Error {
-  constructor(message) {
+  /**
+   * @param {string} message The plugin error message.
+   * @param {Error?} error The original error.
+   */
+  constructor(message, error) {
+    message = `\n${reset.whiteBright.bgRedBright` ${pluginName} `} ${whiteBright(message)}\n`;
+
+    if (error && error.stack) {
+      message += error.stack;
+    }
+
     super(message);
-    this.name = 'LoaderException';
-    this.message = message;
+    this.name = this.constructor.name;
   }
 }
-
-/**
- * @param {string} message The error description.
- * @param {LoaderException|Error|string?} error The original error from catch()
- * @constructor
- */
-const LoaderError = function (message, error = '') {
-  if (error && error instanceof LoaderException) {
-    if (error.toString() === lastError) {
-      // prevent double output same error
-      throw new LoaderException(lastError);
-    }
-    // throw original error to avoid output all nested exceptions
-    lastError = error.toString();
-    throw new Error(lastError);
-  }
-  lastError = message;
-
-  if (error) lastError += `\n\n${redBright`Original Error:`}\n` + error;
-
-  throw new LoaderException(lastError);
-};
 
 /**
  * Return error string as HTML to display the error in browser by HMR.
@@ -57,10 +41,9 @@ const errorToHtml = (error) => {
  * @throws {Error}
  */
 const resolveException = (error, file, templateFile) => {
-  const message =
-    `${pluginHeader} The file ${yellow`'${file}'`} can't be resolved in the template ` + cyan(templateFile);
+  const message = `The file ${yellow`'${file}'`} can't be resolved in the template ` + cyan(templateFile);
 
-  LoaderError(message, error);
+  throw new LoaderException(message, error);
 };
 
 /**
@@ -69,10 +52,10 @@ const resolveException = (error, file, templateFile) => {
  */
 const unsupportedPreprocessorException = (preprocessor) => {
   const message =
-    `${pluginHeader} Unsupported preprocessor '${yellow`${preprocessor}`}'.\n` +
+    `Unsupported preprocessor '${yellow`${preprocessor}`}'.\n` +
     `Please see the possible values in the README on the GitHub.`;
 
-  LoaderError(message);
+  throw new LoaderException(message);
 };
 
 /**
@@ -81,11 +64,11 @@ const unsupportedPreprocessorException = (preprocessor) => {
  */
 const watchPathsException = (dir, paths) => {
   const message =
-    `${pluginHeader} The watch directory not found ${yellow`'${dir}'`}.\n` +
+    `The watch directory not found ${yellow`'${dir}'`}.\n` +
     `Check the ${green`watchFiles.paths`} option:\n` +
     cyan(JSON.stringify(paths, null, '  '));
 
-  LoaderError(message, '');
+  throw new LoaderException(message);
 };
 
 /**
@@ -94,7 +77,7 @@ const watchPathsException = (dir, paths) => {
 const notInitializedPluginError = () => {
   return new LoaderException(
     redBright(
-      `${pluginHeader} Illegal usage of the loader.\n` +
+      `Illegal usage of the loader.\n` +
         `This loader should be used only with the ${green`HtmlBundlerPlugin`} initialized in the ${yellow`Webpack 'plugins' option`}!\n` +
         `See the documentation: ${cyanBright`https://github.com/webdiscus/html-bundler-webpack-plugin`}`
     )
@@ -107,9 +90,7 @@ const notInitializedPluginError = () => {
  * @returns {Error}
  */
 const preprocessorError = (error, file) => {
-  return new LoaderException(
-    `${pluginHeader} Preprocessor failed\nFile: ${cyan(file)}\n` + (error.stack || error.message)
-  );
+  return new LoaderException(`Preprocessor failed.\nFile: ${cyan(file)}`, error);
 };
 
 /**
@@ -118,9 +99,7 @@ const preprocessorError = (error, file) => {
  * @returns {Error}
  */
 const compileError = (error, file) => {
-  return new LoaderException(
-    `${pluginHeader} Template compilation failed\nFile: ${cyan(file)}\n` + (error.stack || error.message)
-  );
+  return new LoaderException(`Template compilation failed.\nFile: ${cyan(file)}`, error);
 };
 
 /**
@@ -129,13 +108,10 @@ const compileError = (error, file) => {
  * @returns {Error}
  */
 const exportError = (error, file) => {
-  return new LoaderException(
-    `${pluginHeader} Export of compiled template failed\nFile: ${cyan(file)}\n` + (error.stack || error.message)
-  );
+  return new LoaderException(`Export of compiled template failed.\nFile: ${cyan(file)}`, error);
 };
 
 module.exports = {
-  LoaderError,
   errorToHtml,
   resolveException,
   unsupportedPreprocessorException,