+++ test/cases/msg-exception-option-entry-dir-not-found/webpack.config.js
@@ -1,0 +1,17 @@
+const path = require('path');
+const HtmlBundlerPlugin = require('../../../');
+
+module.exports = {
+  mode: 'production',
+
+  output: {
+    path: path.join(__dirname, 'dist/'),
+  },
+
+  plugins: [
+    new HtmlBundlerPlugin({
+      // test: directory is not found
+      entry: 'src/pages-invalid/',
+    }),
+  ],
+};