+++ test/cases/msg-exception-option-entry-not-dir/webpack.config.js
@@ -1,0 +1,17 @@
+const path = require('path');
+const HtmlBundlerPlugin = require('../../../');
+
+module.exports = {
+  mode: 'production',
+
+  output: {
+    path: path.join(__dirname, 'dist/'),
+  },
+
+  plugins: [
+    new HtmlBundlerPlugin({
+      // test: entry is not a directory
+      entry: 'src/pages/index.html',
+    }),
+  ],
+};